Restrict shown featured editions to those with translations.
This commit contains a fairly hairy named scope on the
`EditionWorldLocation` join model, so that we can filter those
joins based on which associated `Edition`s have translations. I'd
hoped to be able to `merge` a higher-level scope from Edition, i.e.

    joins(:edition).merge(Edition.with_translations(*locales))

but ActiveRecord complains about there being no `translations`
association on `EditionWorldLocation` (which, indeed, there is not).

Ultimately if we're going to throw `EditionWorldLocation` objects
around in a manner similar to `Editions` (or at least pass them to
presenters in a similar way in the controller), I think it's right
that they have a similar internal API for scoping against
translations. I just wish there was a more concise way to do it.