Add translated worldwide offices.
This follows a very similar form to the translation for world locations, although
some of the admin interface is different.

Firstly, it should be noted that to avoid application downtime this commit doesn't
remove *any* columns from the worldwide_offices table. Once the application is
deployed and restarted, the vestigial content columns on that table should be
removed, but not before.