Revert removal of title, summary and body from edition
Even though we don't have any code relying on the these fields being present,
we think that ActiveRecord is automatically including the field names in the
queries it generates. As such, the removal of these fields still causes a
problem with deployment as old processes will expect the fields to exist.

This commit amends the migration[1] so that it'll be a no-op when run on
production (it's already run everywhere else up to and including staging). It
also adds a migration to re-add the title, summary and body fields on all
machines that have already run the original migration. This is almost
identical to the migration we used last time we were in the mess[2].

[1] e01713decd5846580b6220af62a25f2c7ee2083a
[2] 8b26d249637255f71b0f633d2920c5b2273d116c