Split out Organisation#show into #documents
This should speed up the organisation pages dramatically, as they no
longer load documents on #show.

Additionally we're now paginating the list of organisation editions, to
make the featuring experience faster.

The navigation between tabs should remain the same, as I've co-opted the
tab interface to provide a proper URL instead of a fragment identifier.

https://www.pivotaltracker.com/story/show/44486835