Inject a new WorldwideOffice between WorldwideOrganisation and Contact
The rationale is that we imminently need to add some more information to a contact when it's related to a worldwide organisation, but this information isn't relevant for any other contactable type, so we want a new object type.  WorldwideOffice is the best name because this thing does actually represent a single office.