We don't manage contacts for worldwide orgs, we manage offices
Transition the admin UI over from letting the user manipulate contacts to letting them manipulate offices instead.  Some of the code here is slightly awkward because there's nothing on a worldwide office that needs manipulating yet, and we have to keep referencing office.contact to show things.

There might be scope to introduce delegation of methods from worldwide_office instances down to their contact instances in a future commit, but the contact API is quite wide.