Use proper URLs for localised content
This adds a patch to rails routing adding a `localised` option.
Setting it to true adds a (.:locale) matcher to then end of any
path (before the format), allowing URLs in the following forms to be
matched:

  /documents => {locale: 'en'}
  /documents.fr => {locale: 'fr'}
  /documents.json => {locale: 'en', format: 'json'}
  /documents.fr.json => {locale: 'fr', format: 'json'}

It also sets the default locale for each route to English, so
generating a route such as worldwide_priority_path(locale: 'en') will
return /government/priority (without the locale extension).  N.B.
passing in a locale as a symbol will not work due to a bug in rails
routing somewhere.