Fake GdsApi::Rummager not DocumentFilter
Faking the behaviour of DocumentFilter means that we don't test an
important part of the implementation of our filter pages. This change
means that all of our tests do exercise our real filter
implementation(s). This feels like a better cut-point at which to
introduce the fake behaviour.

Looking forward, it feels as though the fake implementation really
belongs in the same package as Rummageable and GdsApi::Rummager, but
these classes are in different packages from each other at the moment,
so we can't do that yet.

I've also changed NotQuiteAsFakeSearch::Store to not be a singleton, but
rather be passed in the constructor explicitly to the objects which
collaborate with it. I think this explicity dependency-passing style
makes the relationships between objects more clear and explicit.