Present services on the public facing views for an office
Using a ContactPresenter to make WorldwideOffice instances look enough like a Contact that we can continue to use the contacts/_contact partial.  This presenter delegates everything to the contact if it's there, otherwise it delegates to the model as usual.

Have added the display of services to this partial (it's robust enough to work for Contact instances or Contact instances wrapped in a ContactPresenter.

This most definitely needs some styling.