Replace youtube cucumbers with js unit tests
The unit tests ensure that the plugin we have calls out to the nomensa
plugin with the correct params. While this would now miss the case that
we change the class round govspeak I feel this is probably ok. In an
ideal world we would be able to mock the nomensa library in cucumber so
we could have such a test, but without that we are then loading youtube
videos into our cucumber tests.