Convert the edit edition translation link into a button.
I'm planning on adding a delete button to go along with the edit button
and by moving the button "into" the relevant translation section, we
gain extra context for the button and avoid the need to say what it will
allow you to edit.

The markup/styling may not be great, but I think it gets me a bit
further forward.