Allow admin user to remove translation for edition.
The delete button is shown within the section for each translation to
make it clear to which translation it refers.

The delete button is only displayed if the edition itself is deletable.
This means that it will not be possible to delete the translation for a
published edition i.e. similar to the workflow for supporting pages.

I've chosen to add the delete button to the translations tab of the
admin editions show page rather than introducing an equivalent of the
supporting page show page for each translation. Doing it this way is
consistent with the existing edit button and reduces the number of steps
the user has to perform in order to delete a translation.

I've chosen to include the `::Translatable` module into
`Edition::Translatable`. It's possible that some of the methods on the
latter might be moved onto `::Translatable`, i.e. the ones that are
independent of editions.