Pretend common web file extensions don't exist
Some files with a URL ending with `.aspx` will successfully pass the
AttachmentCache validation, and therefore end up on disk, but will then
subsequently not pass the Attachment validation and break the import
they're part of.

This adds a list of common dynamic extensions (PHP, ASP.NET etc.) and
chooses to ignore them when determining the name to save the file
as. This forces the AttachmentCache to use other logic to determine the
filename, which in the case of the attached bug fixes the issue.

https://www.pivotaltracker.com/story/show/44945211