Add functionality for translating News Articles
Making it possible to translate News Articles is trivial (ensuring that
`NewsArticle#translatable?` returns true and that we populate the
`@missing_translations` array in the controller). Unfortunately, testing this
in the same way that we're testing the translation functionality for Worldwide
priorities has lead me to add 11 controller tests, which will definitely slow
the test suite.

My current plan is to add this duplication in the controllers and tests and to
see what I can do about it later.