DRY up all the translation-related controller tests
This follows the pattern we've used previously of introducing a macro like
method that generates all the tests.

As noted in the previous commit, this isn't ideal as it makes it very easy to
rapidly increase the number of tests in the codebase. I'm hoping that I'll be
able to reduce the effect of this in a future commit.