Convert document filter pages to use mustache
This is the first pass at switching the filter lists to use mustache to
render results. The mustache can then be shared on both the server and
the client.

I have slightly modified the `DocumentFilterJsonPresenter` to make it a
generic `DocumentFilterPresenter` which also contains an `as_json`
method. This lets us then use idential hash for both javascript and
ruby.