Simplify the news article and priority controllers
I don't particularly think it's any worse to access
`@edition.missing_translations` in the template rather than assigning it to
`@missing_translations` in the controller. The advantage is that I *think*
this might make it possible to test the translation functionality in the
edition controller tests, rather than having to test it in each of the
controllers that correspond to translatable models.