Introduce GenericEditionsController
My hope is that this controller will make it possible for us to reduce the
number of tests we have. At present, we make use of the macro style tests to
ensure that the different document specific controllers behave as intended.
This is overkill when the behaviour we're testing is provided by the
`EditionsController` or one of the generic templates (those in
app/views/admin/editions). The `GenericEditionsController` should make it
possible to test this common behaviour while reducing our total number of
tests.

The `generic_editions` resource is only defined in the test environment: We
definitely don't want it in production and I don't think it's particularly
useful in development.

I've had to extend the `GenericEdition` object so that we can dictate how it
responds to the `#translatable?` method, as `Edition#translatable?` returns
false by default.

I've moved all the translation related tests from the `should_be_translatable`
macro into the new `Admin::GenericEditionsController::TranslationTest`.

I've created the test/functional/admin/generic_editions_controller_tests
folder so that we can have multiple behaviour related tests using the
`GenericEditionsController`.

I should be able to remove some of the hand-rolled generic_edition path/url
methods in test/factories/editions.rb in a future commit.

I *think* that this new controller will allow us to reduce the number of
Cucumber scenarios if we wish. Instead of having separate scenarios for
translating Worldwide priorities and News articles, for example, we could have
a scenario that talks about Translatable editions.