Change Api::Responder to support links
If links are present in the options they will be dropped into the _response_info json blob, and set as the Link HTTP header.

We use LinkHeader to render, but have a slightly saner API for sending the links.

It wants:
 [[url_1, [[attr_1, url_1_attr_1_value], [attr_2, url_1_attr_2_value]]],
  [url_2, [[attr_1, url_2_attr_1_value], [attr_2, url_2_attr_2_value]]]]

We use:
 [[url_1, {attr_1 => url_1_attr_1_value, attr_2, url_1_attr_2_value}],
  [url_2, {attr_1 => url_2_attr_1_value, attr_2, url_2_attr_2_value}]]

And handle the conversion for you.  URLs can't be a hash (you could specify the same URL for multiple links), but the attrs can.