Remove title, summary and body from editions
This is the third attempt at removing these columns from the editions table.
They were first removed when we added the edition_translations table[1] but
that change had to be reverted[2] when we realised that the removal of the
columns would cause downtime for end users. The second removal attempt[3] came
later but again had to be reverted[4] when we realised that stale ActiveRecord
processes would still be automagically referring to those columns. We
eventually deployed a fix[5] that should mean we can now remove these columns
for good.

[1]: 77f82bb28a061ce8188245a37c5c515d8798d755
[2]: 8b26d249637255f71b0f633d2920c5b2273d116c
[3]: ef5f97dd535a1c3af6cd39a49bc227830b6bf17e
[4]: c967cdd4803096191f7443678e1ae52ef540f0de
[5]: d1fc876aae4e7a8dbb8ca74e4b80452102e1e4e4