Reduce number of tests around showing govspeak
The `should_allow_showing_of` test macro was used in 10 controller tests. The
test in the macro was exercising logic in the 'editions/document_preview'
partial which is shared by all of these controllers. By changing these tests
to use the `GenericEditionsController` we reduce the number of functional
tests by 9 (from 10 to 1).