Use FactoryGirl.build_stubbed instead of a hand-rolled solution.
`FactoryGirl.build_stubbed` takes care of all this (and more).

I had run into problems with the previous implementation where, because
the "stub record" was still an instance of the `ActiveRecord::Base`
subclass (albeit with lots of stubbed method), and because its
`#new_record?` method returned `true`, it was possible for a database
query to be initiated by something as simple as inspecting the record or
calling an association method.

This problem was compounded in the cases I saw, because Mocha tried to
report the unexpected invocation of `select` on the `ActiveRecord`
connection, but in doing so it also wanted to list the stubbed methods
on the record. The latter caused a recursive call to the record's
`#inspect` method and an eventual stack overflow error.