Reduce number of tests around audit trail
The `should_show_document_audit_trail_for` test macro was used in 9 controller
tests. The test in the macro was exercising logic in the
'editions/audit_trail_entry' partial which is shared by all of these
controllers. By changing these tests to use the `GenericEditionsController` we
reduce the number of functional tests by 16 (from 18 to 2).

*NOTE* I've had to create an empty 'generic_editions/_form' partial so that we
can render 'editions/edit' correctly.