Reduce number of tests around revising
The `should_allow_revision_of` test macro was used in 7 controller tests. The
tests in the macro were exercising logic in the 'editions/navigation' partial
which is shared by all of these controllers. By changing these tests to use
the `GenericEditionsController` we can reduce the number of functional tests
by 18 (from 21 to 3).