Reduce number of tests around deleting
The `should_be_able_to_delete_an_edition` test macro was used in 10 controller
tests. The tests in the macro were exercising logic in the
'editions/navigation' partial and the editions controller, all of which are
shared by these controllers. By changing these tests to use the
`GenericEditionsController` we can reduce the number of functional tests by 63
(from 70 to 7).

*NOTE*. It should be possible to further simplify the tests for the
presence/absence of the delete button as the template only switches based on
the value of `Edition#deletable?` (i.e. it doesn't care about the actual state
of the edition).