Allow admin users to create & edit translated person biographies.
This includes the ability to edit right-to-left translations in the
admin UI.

There is quite a bit of duplication between the
`PersonTranslationsController` and other similar controllers, but I will
hope to address this in a later commit.

Despite the fix in 13237e4ea7e54c4d3bc9bd8d244fe08ff0655b2f, there's
still some awkwardness in the setup of `PersonPresenterTest`. This seems
to be because the FactoryGirl stub is still affected by Globalize3 and
therefore attempts are made to query the `person_translations` table
unless the `#translations` method is stubbed.