Reduce number of tests around rejecting
The `should_be_rejectable` test macro was used in 10 controller tests. The
tests in that macro were exercising the 'editions/navigation' partial, the
'editions/rejected_by' partial, and the 'editions/audit_trail_entry' partial;
all of which are shared by those controllers. By changing these tests to use
the `GenericEditionsController` we can reduce the number of functional tests
by 45 (from 50 to 5).