Ensure only one json next link exists on the page
Due to the way slimmer moves the next link to the header we were ending
up with multiple links on the page. This ensures that one is removed
before adding the next one.

The new one should really be added to the head of the page but this will
do for now.