Make translated person biographies available on public website.
Translations for right-to-left languages appear to be displayed
correctly i.e. they use the right-to-left styles. This came "for free"
because of previous work by @tomafro & @edds.

I have followed the pattern used elsewhere of having a couple of
end-to-end Cucumber tests which cover both adding a translation *and*
viewing it on the public website.

The tables I have used in a couple of steps are somewhat overkill, but
the follow the pattern established elsewhere and mean that it would be
easier to add other translated attributes in the future.

I've also added an assertion to a functional test to force me to
introduce the link from the admin UI to the translated person page on
the public website.