Make role name & responsibilities translatable.
This adds a "Manage translations" link for each role in the admin UI
allowing the name & responsibilities a role to be translated into
multiple languages. This means it's a relatively big commit, but I'm
comfortable with this because most of the functionality is very similar
to other model translations. I'll explain some of the differences below.

The translated name & responsibilities will show up on the
locale-specific version of the public-facing person page, as long
as both the person and *all* their current roles have translations into
a given locale.

I've introduced a couple of methods into the `PersonPresenter` which
cope with the logic described above, but it feels as if it might make
sense to introduce some kind of aggregate class which copes with the
combination of multiple translatable models appearing on a single page.

Given that we also display names of previous roles, we might also
want to take these into account when deciding whether we can display
a person page for a given locale, but I haven't tackled this here.

The admin UI for roles and people was already pretty clunky and I think
these changes may make that more obvious, but that feels like a bigger
issue which I have not tackled here.

As elsewhere I have introduced an *end-to-end* Cucumber scenario which
adds a translation to an existing role and checks that the translated
content is available on the public-facing person page. I have only added
a test for adding a new translation and not editing an existing one,
because I feel that that functionality is covered by the functional
tests.

It's worth noting that I had to use the `as: :role` option in the form
helper in `views/admin/role_translations/edit.html.erb`, because the
role is always an instance of a subclass of `Role`. Without the option
the input fields would be namespaced according to the subclass which
would mess up the parsing of params in `RoleTranslationsController`.

One small thing I had to change was in `Admin::RolesController#destroy`
where it turns out that the translations associated with a role are
destroyed when the role itself is destroyed, which in turn means that
the translated attributes are no longer available.

I've changed all the places that were using
`ModelStubbingHelpers#stub_record` to build a role, to use
`#stub_translatable_record` instead in order to avoid the tests hitting
the database and failing the expectation set by
`#disable_database_queries`.