Only allow translations for worldwide roles.
This means that the "Manage translations" link only shows up against
worldwide roles in the admin UI. I feel that this is likely to make
things less confusing for admins who are not interested in worldwide
roles.

However, if we end up deciding that we want to show translated versions
of previous roles on the public-facing person page, I can see that we
might want to relax this i.e. if an ambassador was formerly in a
non-worldwide role.