Fix logic in TranslatableModel#available_in_multiple_languages?
I've added unit tests for this module. The third one demonstrates the
flaw in the logic:

If a model has translated attributes, but none of those translated
attributes are *required* attributes, then no Globalize3 translations
will be created and `#translated_locales` will return `[]`.

This is probably a slightly unusual situation, but it is possible e.g. a
person is created with *no* English biography. A Spanish biography is
then added to this person, but
`TranslatableModel#available_in_multiple_languages?` would return false
and it would not be possible to view the Spanish version of the person
page.