Use tmp folder for temporary test files
This makes the tests use the tmp folder to put the files for the bulk
upload test into. Previously if the tests didn't finish for what ever
reason it would leave a folder in the root folder. This makes it only
leave that folder in the tmp folder so its not in your way.