Make all roles/people fetched in OrganisationsController unique
Follow the pattern established by board_members of using a RolesPresenter and retrieving unique_people from it.  For some reason only the ministers were being reduced to the filled roles so I've replicated that behaviour, but I think that iterating over the unique people has the same effect on the others.