Add localised routing to people
This ensures that /people/<person-slug>.<locale> works correctly. Previously,
you would've had to add the '?locale=<locale>' querystring to view a person in
a different locale. This also ensures that links to people take you to the
page in the locale you're viewing the current page in.

*NOTE* I've had to extend the `stub_translatable_record` helper so that the
`#translations` method responds with something that can answer to `#where`,
which in turn responds with something that can answer to `#exists?`. This is
because methods (specifically, `#available_in_locale?`) now comes into play
when generating the route for a person.