Add organisations as tags in detailed guide api response
In order to simplify some of the reporting logic the ops team are
developing around our content it would be useful for the detailed
guides api to expose details of the organisations responsible.

I know it's controversial to represent these as tags, but it's the
simplest representation and I believe it works so long as we
recognise that the tag form is only one representation of an
organisation in our system.