Avoid "Edit" link collision: Rename "Eddie Editor" to "Eddie Depteditor"
In our cucumber if we're logged in as an editor and we click_link "Edit" the first link on the page with the text "Edit" is our name and this takes us to the wrong place.  It's easier to rename our editor than it is to introduce extra styling just to help the tests target the document nav actions.