Tidy up Uploader-related code
I performed an audit of the import/uploader related code and found
that there were a lot of tests claiming to be "fast tests" when in
fact they were using factories and hitting the database. I've gone
through and update the tests accordingly, and also tidied up the
module structure so that everything is autoloaded, removing the need
for lots of explicit requires.