Ensured no duplicate people on ministerial list
I've used Tekin's RolesPresenter which contains helpers to ensure we
have a unique group of people. This class is really a domain class
rather than forming part of the presentation layer: it's a good
representation of a sub-group of roles for an organisation.

This has radically simplified the `ministers_by_organisation` method in
the `MinisterialRolesController` and removed the need for the select
workaround.