Move worldwide priority/edition joins to document
Previously the worldwide priorities were joining to edition ids, which
is incorrect: if the document is subsequently republished, then we won't
be pointing to the right edition. This change appropriates the existing
join for policies and uses it for priorities also.

The migration is destructive (all existing links between priorities and
editions are removed), but because the code isn't live yet, this
shouldn't be an issue.

The related_documents method here actually is only needed in the tests,
and should probably be a property of the edition factories.