Ensure that policies and priorties don't interfere
When associating policies and priorities to the same edition we
previously had clashes.

Turns out it's not actually possibly yet to associate policies
and worldwide priorities, but this may come up in the future and it's
good to have a central place that we record all edition -> document
relations that works as advertised rather than inconsistent behaviour.

At present the implementation of `related_policy_ids=` and
`worldwide_priority_ids=` isn't fully satisfactory: potentially this
could be refactored into the shared base class using a helper, rather
like `attachable`.