Add tracking to the number of infinte pages shown
To let us know how many people are using the infinite scroll. It should
also let us tune the number of results to return on the first page.

I have put the event namespaced to my name as this was recomended by the
analytics team so they know its mine. I am only planing on running this
for a short period of time (enough to give us a trend) before tracking
something else so I have kept the event name firaly generic.