Enable locale I18n fallback in test mode
Without fallback turned on, any keys that are missing in the current
locale raises an error. This is problematic if you are trying to test
the translation of something and the locale under test does not have
translations for all the keys. It turns out what we really care about
is always having a value for any given key, i.e. if fallback is on,
the default locale must have a value for all keys. So as well as
turning on fallback in test mode, I've added a couple of tests that:
1. ensure all keys in the default locale have a value; 2. all other
locale files are newer than the default locale, i.e. they have been
regenerated to add the new keys.