Correct force publishing counts when multiple old urls
If an imported document had multiple old urls this would cause the
documents/editions to get counted twice.

There's a known (fixed) bug in active record that count does not respect
uniq:

https://github.com/rails/rails/issues/6865

This bugfix will be included in 3.2.13 so if we move to this or a later
version we could clean up this code.