Change Searchable to respect :only param in update_in_search_index
Previously searchable_options[:only] was only used to scope the objects retrieved when using Class.search_index to index all the instances.  This meant if we only wanted to index some instances we had to deal with it in our own callbacks for normal lifecycle events.  Now the only param is used as a guard (by adding .find_by_id(self.id).present?) inside update_in_search_index so that if the instance is not present in the scope it won't be added to the search index.