Add pagination to policy activity pages
I have tried to keep the test for this fairly simple and only assert
that we are exciting the pagination gem we use. That we pass it the
correct page from the query string and that the pagination links exist.
There seems little point to extensivly test the pagination as the we
should assume the gem is tested and works.

Also updated the styling of the pagination slightly to be less
'detailed-guide' as we never use it for detailed guides any more.