Add observer on org to add/remove corp info from search
It observes updates on organisations and adds corp info pages to the index if the org is changing from govuk_status ??? to 'live'.  It will also remove them from the index if the org is changing from govuk_status 'live' to ???.  Create & destroy events on the corp info page should be handled by the default searchable index_after/unindex_after so we don't need to do them here.