Remove user 553 and change any refs to 648
We first change all references then remove the user record. On a
reasonably recent database dump (two days ago) there were no references
to this user in any of the tables, but left in just in case.

Ran before/after full diffs of the database, and it appears to only
remove the record and bump the migration count.