First cut at a hand-rolled auth system
We get an Enforcer instance for an actor (usually an User instance) and a subject (usually an instance of some AR object, but it could be anything).  We then ask it if we can? perform an action.

In turn the Enforcer looks up a series of rules based on the class of the subject (e.g. an Edition might use EditionRules) - with a fall through to ObjectRules which is very pessimistic and always says no.  It then uses the rules to ask if the action can be performed.

We can then use ruby to decide if the actor can perform that action on the subject.

It's not super clear to me if this approach is good.  It's kinda like cancan or the other role/auth libraries out there, but I think it'll give us more flexibility in features (e.g. we want to get reasons for why some actions are denied).