Ensure mustache templates work as expected
There is a slight difference in the way the mustache tempaltes are
rendered on the server and the client. On the server you can use a
`{{.}}` to print out the current object. However, on the client this
wasn't working. Swapping it out for the name of the attribute fixes it
on both the client and the server.

Added some tests so that we ensure that this continues to work on the
client.

https://www.pivotaltracker.com/story/show/46143349