Specialised jenkins script for branches
Added a new script to be run for branches. This script drops and reloads
the schema each time to the fact that different branches might contain
different migrations.

Caveat: The migrations on the branches themselves aren't being run
currently, which potentially is an area for improvement. Potentially the
schema from the master:branch base point should be run and db:migrate
run afterwards.

Caveat: we're using the whitehall_test1 database for this job, which is
also used by the parallel build (currently disabled.) Potentially we
should move it to its own whitehall_test_branches database.