Separate edition workflow from permission, and add full set of actions
Rules for editions have changed a bit to try to push workflow stuff (e.g. you can't force publish something that's not been submitted) back down to the models, where for better or worse, the rest of the business rules live.  We also make sure that if you can't see it (e.g. access limiting) then you can't do anything else to it.