Add dynamic counts of consultations to get involved
There are two incomplete things here:

1. we don't get the count of the responses published in the past 12
months. That's a bit tricky to do because the 'publication date' of a
response is either the explicit published_on field or else the first
created date of one of the attachments. Not only does this make it
difficult to filter the responses by published date, but it also seems
slightly wrong logic (you might prepare the draft with the response
documents some days before actually publishing)

2. similarly the list of recent responses is slightly incorrect. It
returns the 3 most recently closed consultations that have responses,
rather than the 3 most recent responses.