Kill off duplciate imported child publications
Because of a quirk in the way home offce publications are structured
and the way they were scraped, many duplicates have ended up in the
system. This story will clean up these duplicates by deleting both the
document and the source of these duplicates, as identified by the CSV
file. An output CSV is generated so that we have a log of exactly what
is deleted, just in case.