authority_test_helper can run tests inside test:units
If we detect that Rails is defined, we produce a slightly different AuthorityTestHelper module for mixing into the auth tests and we just require the normal test_helper.  If it's not been defined we load fast_test_helper and set up some other lighter objects.

This means running the test within rake test:units won't break because of class mismatches or whatever.