Add jQuery UI slider for media player
The media player relies on the jQuery UI slider. It used to be available
as part of the javascript we got from static. However, with the move to
the header footer only javascript we no longer include that.

This adds the minimum required jQuery UI to get the slider working. I
have also removed the ba-hashchange.js file as we are no longer using
hash change events.

One slight problem I had was that the tests include all javascript files
to run. However, upon trying to load jquery ui twice (once in the
admin.js and once in application.js) it crashed out. I have changed
all.js to only include defined JS.