Rejig routing for admin worldwide offices
Defining the resource as :offices was causing me no end of trouble
because the url helpers that are generated break with the standard
rails conventions. Trying to nest a polymorphic singular resource
under it quickly turned into a nightmare, mainly because you can no
longer rely on rails to just do the right thing when it automatically
generates urls in views and for forms. Using the path: option achieves
the same effect (a neater url), but keeps the routing conventional.