Org editors and writers should be allowed to manage world documents
This *might* change back again once world is live and we can settle down the permissions properly.  For now, we don't want to disrupt the production of content ahead of the launch so we just allow it.  This does mean that the WorldEditionRules is almost redundant, but that might change.  Of course, if it doesn't then we should remove it and simplify.