Users can create statistical data sets again
The issue was that in the #new action we were checking permission
against the new unsaved instance, which didn't have an organisation set.
As statistical data sets are access limited by default, if there is no
organisation no one is able to access the new object, which means that
the 'new' form cannot be displayed.